+++ main.go
@@ -1,55 +1,20 @@
 package main
 
 import (
-	"os"
-
 	"github.com/alecthomas/kong"
 
 	"github.com/bugsnag/bugsnag-cli/pkg/build"
 	"github.com/bugsnag/bugsnag-cli/pkg/log"
+	"github.com/bugsnag/bugsnag-cli/pkg/options"
 	"github.com/bugsnag/bugsnag-cli/pkg/upload"
 	"github.com/bugsnag/bugsnag-cli/pkg/utils"
+	"os"
 )
 
-var package_version = "2.0.0"
-
-// Global CLI options
-type Globals struct {
-	UploadAPIRootUrl  string            `help:"Bugsnag On-Premise upload server URL. Can contain port number" default:"https://upload.bugsnag.com"`
-	BuildApiRootUrl   string            `help:"Bugsnag On-Premise build server URL. Can contain port number" default:"https://build.bugsnag.com"`
-	Port              int               `help:"Port number for the upload server" default:"443"`
-	ApiKey            string            `help:"(required) Bugsnag integration API key for this application"`
-	FailOnUploadError bool              `help:"Stops the upload when a mapping file fails to upload to Bugsnag successfully" default:"false"`
-	Version           utils.VersionFlag `name:"version" help:"Print version information and quit"`
-	DryRun            bool              `help:"Validate but do not process"`
-}
-
-// Unique CLI options
-type CLI struct {
-	Globals
-
-	Upload struct {
-		// shared options
-		Overwrite bool `help:"Whether to overwrite any existing symbol file with a matching ID"`
-		Timeout   int  `help:"Number of seconds to wait before failing an upload request" default:"300"`
-		Retries   int  `help:"Number of retry attempts before failing an upload request" default:"0"`
-
-		// required options
-		AndroidAab         upload.AndroidAabMapping      `cmd:"" help:"Process and upload application bundle files for Android"`
-		All                upload.DiscoverAndUploadAny   `cmd:"" help:"Upload any symbol/mapping files"`
-		AndroidNdk         upload.AndroidNdkMapping      `cmd:"" help:"Process and upload Proguard mapping files for Android"`
-		AndroidProguard    upload.AndroidProguardMapping `cmd:"" help:"Process and upload NDK symbol files for Android"`
-		DartSymbol         upload.DartSymbolOptions      `cmd:"" help:"Process and upload symbol files for Flutter" name:"dart"`
-		ReactNativeAndroid upload.ReactNativeAndroid     `cmd:"" help:"Upload source maps for React Native Android"`
-		ReactNativeIos     upload.ReactNativeIos         `cmd:"" help:"Upload source maps for React Native iOS"`
-		UnityAndroid       upload.UnityAndroid           `cmd:"" help:"Upload Android mappings and NDK symbol files from Unity projects"`
-	} `cmd:"" help:"Upload symbol/mapping files"`
-	CreateBuild          build.CreateBuild          `cmd:"" help:"Provide extra information whenever you build, release, or deploy your application"`
-	CreateAndroidBuildId build.CreateAndroidBuildId `cmd:"" help:"Generate a reproducible Build ID from .dex files"`
-}
+var package_version = "2.1.0"
 
 func main() {
-	commands := CLI{}
+	commands := options.CLI{}
 
 	// If running without any extra arguments, default to the --help flag
 	// https://github.com/alecthomas/kong/issues/33#issuecomment-1207365879
@@ -81,7 +46,7 @@ func main() {
 	case "upload all <path>":
 
 		if commands.ApiKey == "" {
-			log.Error("no API key provided", 1)
+			log.Error("missing api key, please specify using `--api-key`", 1)
 		}
 
 		err := upload.All(
@@ -172,7 +137,7 @@ func main() {
 	case "upload dart <path>":
 
 		if commands.ApiKey == "" {
-			log.Error("no API key provided", 1)
+			log.Error("missing api key, please specify using `--api-key`", 1)
 		}
 
 		err := upload.Dart(commands.Upload.DartSymbol.Path,
@@ -247,7 +212,7 @@ func main() {
 	case "upload unity-android <path>":
 
 		if commands.ApiKey == "" {
-			log.Error("no API key provided", 1)
+			log.Error("missing api key, please specify using `--api-key`", 1)
 		}
 
 		err := upload.ProcessUnityAndroid(
@@ -272,37 +237,31 @@ func main() {
 		}
 
 	case "create-build", "create-build <path>":
+		// Create Build Info
+		CreateBuildOptions, err := build.GatherBuildInfo(commands)
 
-		if commands.ApiKey == "" {
-			log.Error("no API key provided", 1)
+		if err != nil {
+			log.Error(err.Error(), 1)
 		}
 
-		// Build connection URI
-		endpoint, err := utils.BuildEndpointUrl(commands.BuildApiRootUrl, commands.Port)
+		// Validate Build Info
+		err = CreateBuildOptions.Validate()
+
+		if err != nil {
+			log.Error(err.Error(), 1)
+		}
+
+		// Get Endpoint URL
+		endpoint, err = utils.BuildEndpointUrl(commands.BuildApiRootUrl, commands.Port)
 
 		if err != nil {
 			log.Error("Failed to build upload url: "+err.Error(), 1)
 		}
 
-		log.Info("Creating build on: " + endpoint)
+		err = build.ProcessCreateBuild(CreateBuildOptions, endpoint, commands.DryRun, commands.CreateBuild.Timeout)
 
-		buildUploadError := build.ProcessBuildRequest(commands.ApiKey,
-			commands.CreateBuild.BuilderName,
-			commands.CreateBuild.ReleaseStage,
-			commands.CreateBuild.Provider,
-			commands.CreateBuild.Repository,
-			commands.CreateBuild.Revision,
-			commands.CreateBuild.VersionName,
-			commands.CreateBuild.VersionCode,
-			commands.CreateBuild.BundleVersion,
-			commands.CreateBuild.Metadata,
-			commands.CreateBuild.Path,
-			endpoint,
-			commands.DryRun,
-		)
-
-		if buildUploadError != nil {
-			log.Error(buildUploadError.Error(), 1)
+		if err != nil {
+			log.Error(err.Error(), 1)
 		}
 
 		log.Success("Build created")